ASDF Refactoring and Unit Tests
* Moved FederatedASDFDataSetMemVariant to legacy folder as the DBVariant has proven
  robust for both small and large ASDF files over a range of use-cases
* Renamed FederatedASDFDataSetDBVariant to _FederatedASDFDataSetImpl, which now contains
  the implementational details of FederatedASDFDataSet
* Removed merging functionality altogether -- clients now deal with merging returned streams
  depending on use-case
* Made fetching of traces from the underlying ASDF file more robust -- FederatedASDFDataSet
  now takes an additional parameter in the constructor, which controls the size of the
  read-buffer for single reads
* Added detailed tests for FederatedASDFDataSet
